cron_sms/requirements.txt
python-dotenv==0.1.5
deta==1.1.0
twilio==6.29.1
